refactor: refactor footer
