--- components/common/Footer.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-  return <div></div>;
-};
-
-export default Footer;
