+++ pages/booktalk/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Footer from '../../components/common/Footer/Footer';
+import Layout from '../../components/common/Layout';
+
+const index = () => {
+  return (
+    <Layout>
+      booktalk
+      <Footer />
+    </Layout>
+  );
+};
+
+export default index;
