+++ pages/home/index.tsx
@@ -18,10 +18,6 @@ import {
   PinIcon,
   HeartIcon,
   HomeLightMoreIcon,
-  NavBookGrayIcon,
-  NavHomeColorIcon,
-  NavPersonGrayIcon,
-  NavPinGrayIcon,
   InterestingRegionMoreIcon,
   Ranking1Icon,
   Ranking2Icon,
@@ -36,6 +32,7 @@ import {
 import { uesFetchMyInfo } from '../../hooks/queries/mypage';
 import { NewLogoWhite } from '../../assets/img';
 import RenderWhen from '../../components/common/RenderWhen';
+import Footer from '../../components/common/Footer/Footer';
 
 function Home() {
   const router = useRouter();
@@ -256,60 +253,7 @@ function Home() {
       <HotBookTalkSliderWrapper>
         <HotBookTalkSlider data={data?.booktalkDeadlineUpcoming} />
       </HotBookTalkSliderWrapper>
-      <FooterWrapper>
-        <Footer>
-          <IconsWrapper>
-            <IconWrapper>
-              <Image src={NavHomeColorIcon} alt="홈 화면 바로가기 아이콘" />
-              <IconText>홈</IconText>
-            </IconWrapper>
-            <IconWrapper>
-              <Image
-                src={NavPinGrayIcon}
-                alt="지역 화면 바로가기 아이콘"
-                onClick={() => {
-                  if (data?.myCityBooktalkCount === null) {
-                    router.push('/booktalk/search/UIJEONGBU_SI');
-                  } else {
-                    router.push(`/booktalk/search/${myInfo?.city}`);
-                  }
-                }}
-              />
-              <UnClickedIconText>지역</UnClickedIconText>
-            </IconWrapper>
-            <IconWrapper>
-              <Image
-                src={NavBookGrayIcon}
-                alt="소피스토리 화면 바로가기 아이콘"
-                onClick={() => {
-                  if (data?.name === null) {
-                    alert('비회원은 접근할 수 없는 페이지예요 :(');
-                    router.push('/auth');
-                  } else {
-                    router.push('/sophyStory');
-                  }
-                }}
-              />
-              <UnClickedIconText>소피스토리</UnClickedIconText>
-            </IconWrapper>
-            <IconWrapper>
-              <Image
-                src={NavPersonGrayIcon}
-                alt="MY 페이지 바로가기 아이콘"
-                onClick={() => {
-                  if (data?.name === null) {
-                    alert('비회원은 접근할 수 없는 페이지예요 :(');
-                    router.push('/auth');
-                  } else {
-                    router.push('/mypage');
-                  }
-                }}
-              />
-              <UnClickedIconText>나의 소피</UnClickedIconText>
-            </IconWrapper>
-          </IconsWrapper>
-        </Footer>
-      </FooterWrapper>
+      <Footer />
 
       {data?.is_author && data?.booktalkCount !== null && (
         <OpenBookTalkButton onClick={() => router.push('/author/region')}>
@@ -580,59 +524,6 @@ const MoreHotBookTalk = styled.div`
 
 const HotBookTalkSliderWrapper = styled.div`
   margin-bottom: 9rem;
-`;
-
-const Footer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-
-  width: 37.5rem;
-  height: 8.3rem;
-
-  padding-left: 3rem;
-  padding-right: 3rem;
-
-  margin-top: 1rem;
-
-  background-color: ${theme.colors.white};
-
-  box-shadow: 0rem -0.4rem 0.8rem rgba(54, 57, 60, 4%);
-`;
-
-const FooterWrapper = styled.div`
-  position: fixed;
-  bottom: 0;
-  overflow-x: hidden;
-  display: flex;
-  justify-content: center;
-`;
-
-const IconWrapper = styled.div`
-  text-align: center;
-  width: 4.9rem;
-  height: 5.1rem;
-
-  cursor: pointer;
-`;
-
-const IconText = styled.div`
-  color: ${theme.colors.green06};
-  ${theme.fonts.caption};
-  text-align: center;
-`;
-
-const UnClickedIconText = styled.div`
-  color: ${theme.colors.gray06};
-  ${theme.fonts.caption};
-  text-align: center;
-`;
-
-const IconsWrapper = styled.div`
-  width: 32.5rem;
-  display: flex;
-  justify-content: space-between;
 `;
 
 const OpenBookTalkButton = styled.button`
