+++ pages/mypage/index.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import MyProfile from '../../components/mypage/MyProfile';
+import Footer from '../../components/common/Footer/Footer';
 
 const Mypage = () => {
   return (
     <div>
       <MyProfile />
+      <Footer />
     </div>
   );
 };
